lseek, flock y access: tres argumentos más leídos con el ancho equivocado
Misma forma que el PR #1309, en otras tres llamadas. El syscall declara el
argumento más ancho de lo que este árbol lo leía, así que un valor malo no
daba error: daba una respuesta plausible y equivocada.

`lseek` recibía `whence` como `u8`. El byte bajo de 0x100 es 0, o sea
SEEK_SET, de modo que `lseek(fd, off, 0x100)` movía la posición a `off` en vez
de contestar EINVAL y el que llamaba leía otros bytes sin enterarse. Por el
mismo camino, SEEK_DATA (3) y SEEK_HOLE (4) eran indistinguibles de no pedir
nada. Y con SEEK_SET un desplazamiento negativo se volvía una posición cerca
de 2^64, que sale del syscall pareciendo un errno; `vfs_setpos()` contesta
EINVAL. Para SEEK_CUR y SEEK_END el negativo es legal —son relativos— y
Linux comprueba la posición resultante, no el argumento.

`flock` leía la orden por un `bitflags` de 8 bits, y eso aceptaba tres cosas
que Linux rechaza, las tres contestando "cerrojo puesto": `flock(fd, 0)`, que
no nombra ninguna orden; `flock(fd, LOCK_SH | LOCK_EX)`, que nombra dos; y
cualquier bit alto, porque 0x102 tiene byte bajo 2 y se tomaba por LOCK_EX.
`flock_translate_cmd()` es un `switch` sobre tres valores exactos, después de
quitar LOCK_NB, sobre un `unsigned int`. El cerrojo sigue sin tomarse, como
antes; lo que no puede hacer es dar éxito a lo que el kernel rechaza, porque
así es como se comprueba si la función existe.

`faccessat` enmascaraba el modo con `& 0o7` donde `do_faccessat()` valida:
`if (mode & ~S_IRWXO) return -EINVAL;`. Enmascarado, `access(path, 8)`
preguntaba F_OK y contestaba "sí, puedes" para cualquier fichero que
simplemente exista, que es la única respuesta que una sonda de permisos no
puede inventarse.

12 tests nuevos, validados con once mutaciones que caen las once con nombre.

+++ linux-syscall/src/bsd/mod.rs
@@ -141,7 +141,7 @@ impl Syscall<'_> {
             }
             sys::CLOSE => BsdRet::from_result(self.sys_close(a0.into())),
             sys::CLOSE_RANGE => BsdRet::from_result(self.sys_close_range(a0, a1, a2)),
-            sys::LSEEK => BsdRet::from_result(self.sys_lseek(a0.into(), a1 as i64, a2 as u8)),
+            sys::LSEEK => BsdRet::from_result(self.sys_lseek(a0.into(), a1 as i64, a2)),
             sys::FSYNC => BsdRet::from_result(self.sys_fsync(a0.into())),
             sys::FDATASYNC => BsdRet::from_result(self.sys_fdatasync(a0.into())),
             sys::TRUNCATE => BsdRet::from_result(self.sys_truncate(a0.into(), a1)),
